Don't save "draw" and "pick" values in the database.
This was vestigal code from a prior implementation, for dealing with
special-case cards where those values could not be implied from the
text itself. However, in looking over the real CAH sets, the only card
that needed explicit settings was "What's the next superhero/sidekick
duo?", which is a pick-2 without blanks.

Since I don't have a way for users to specify draw/pick counts, I see
it unnecessary to have to deal with this case.

We now compute pick count at time-of-use rather than at time-of-save.

Signed-off-by: Brandon Streiff <git@streiff.net>